requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
SQLAlchemy==1.4.52
typing_extensions==4.10.0
urllib3==2.2.1

python-dateutil~=2.9.0.post0